+++ CHANGELOG.md
@@ -17,6 +17,12 @@ changes in stylesheets:
 - use relative fontsize for literal, such that it may appears in title, headings, & list
 - make hyperlink to footnote superscripts
 
+chagne re CLI:
+
+- change publisher version appending logic, append a file
+- add option -m
+- remove option -e
+